refactor(agents): unify and refactor AgentCard component with marketplace and studio views
- Create a single AgentCard component supporting both 'marketplace' and 'studio' contexts
- Implement MarketplaceListView and MarketplaceGridView for marketplace context display
- Implement StudioGridView with deployment status and metrics for studio context
- Add KYABadge, TrustScore, RatingStars, and PriceBadge subcomponents for modular UI parts
- Replace previous studio AgentCard with deprecated re-export for backward compatibility
- Update imports in marketplace page and other components to use new AgentCard location
- Fix sub-category button key and label rendering in marketplace page for better consistency
